requirements.txt
streamlit>=1.28.0
slack-sdk>=3.21.0
openai>=1.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
typing-extensions>=4.5.0
requests>=2.31.0
tiktoken>=0.5.2